Removing DRCSimulationStarter, AvatarSimulationFactory and AvatarSimulation
