--- ihmc-avatar-interfaces/src/main/java/us/ihmc/avatar/DRCFlatGroundWalkingTrack.java
@@ -1,209 +1,0 @@
-package us.ihmc.avatar;
-
-import us.ihmc.avatar.drcRobot.DRCRobotModel;
-import us.ihmc.avatar.factory.AvatarSimulation;
-import us.ihmc.avatar.factory.AvatarSimulationFactory;
-import us.ihmc.avatar.initialSetup.DRCGuiInitialSetup;
-import us.ihmc.avatar.initialSetup.DRCSCSInitialSetup;
-import us.ihmc.avatar.initialSetup.RobotInitialSetup;
-import us.ihmc.commonWalkingControlModules.capturePoint.splitFractionCalculation.SplitFractionCalculatorParametersReadOnly;
-import us.ihmc.commonWalkingControlModules.configurations.HighLevelControllerParameters;
-import us.ihmc.commonWalkingControlModules.configurations.WalkingControllerParameters;
-import us.ihmc.commonWalkingControlModules.desiredFootStep.footstepGenerator.FootstepAdjustment;
-import us.ihmc.commonWalkingControlModules.desiredFootStep.footstepGenerator.HeadingAndVelocityEvaluationScriptParameters;
-import us.ihmc.commonWalkingControlModules.desiredFootStep.footstepGenerator.HeightMapBasedFootstepAdjustment;
-import us.ihmc.commonWalkingControlModules.dynamicPlanning.bipedPlanning.CoPTrajectoryParameters;
-import us.ihmc.commonWalkingControlModules.highLevelHumanoidControl.factories.ContactableBodiesFactory;
-import us.ihmc.commonWalkingControlModules.highLevelHumanoidControl.factories.HighLevelControllerStateFactory;
-import us.ihmc.commonWalkingControlModules.highLevelHumanoidControl.factories.HighLevelHumanoidControllerFactory;
-import us.ihmc.commonWalkingControlModules.highLevelHumanoidControl.factories.WalkingProvider;
-import us.ihmc.euclid.transform.RigidBodyTransform;
-import us.ihmc.humanoidRobotics.communication.packets.dataobjects.HighLevelControllerName;
-import us.ihmc.humanoidRobotics.communication.subscribers.PelvisPoseCorrectionCommunicatorInterface;
-import us.ihmc.robotics.controllers.ControllerFailureListener;
-import us.ihmc.robotics.robotSide.RobotSide;
-import us.ihmc.robotics.robotSide.SideDependentList;
-import us.ihmc.ros2.ROS2NodeBuilder;
-import us.ihmc.ros2.RealtimeROS2Node;
-import us.ihmc.sensorProcessing.parameters.HumanoidRobotSensorInformation;
-import us.ihmc.simulationConstructionSetTools.util.HumanoidFloatingRootJointRobot;
-import us.ihmc.simulationconstructionset.SimulationConstructionSet;
-import us.ihmc.wholeBodyController.RobotContactPointParameters;
-import us.ihmc.yoVariables.variable.YoBoolean;
-
-import javax.swing.*;
-import java.util.ArrayList;
-
-import static us.ihmc.humanoidRobotics.communication.packets.dataobjects.HighLevelControllerName.DO_NOTHING_BEHAVIOR;
-import static us.ihmc.humanoidRobotics.communication.packets.dataobjects.HighLevelControllerName.WALKING;
-
-public class DRCFlatGroundWalkingTrack
-{
-   // looking for CREATE_YOVARIABLE_WALKING_PROVIDERS ?  use the second constructor and pass in WalkingProvider = YOVARIABLE_PROVIDER
-
-   private final AvatarSimulation avatarSimulation;
-
-   private final RealtimeROS2Node realtimeROS2Node = new ROS2NodeBuilder().buildRealtime("flat_ground_walking_track_simulation");
-
-   private static boolean createYoVariableServer = System.getProperty("create.yovariable.server") != null
-         && Boolean.parseBoolean(System.getProperty("create.yovariable.server"));
-
-   public DRCFlatGroundWalkingTrack(RobotInitialSetup<HumanoidFloatingRootJointRobot> robotInitialSetup, DRCGuiInitialSetup guiInitialSetup, DRCSCSInitialSetup scsInitialSetup,
-                                    boolean useVelocityAndHeadingScript, boolean cheatWithGroundHeightAtForFootstep, DRCRobotModel model)
-   {
-      this(robotInitialSetup, guiInitialSetup, scsInitialSetup, useVelocityAndHeadingScript, cheatWithGroundHeightAtForFootstep, model,
-           WalkingProvider.VELOCITY_HEADING_COMPONENT, new HeadingAndVelocityEvaluationScriptParameters(), null, null); // should always be committed as VELOCITY_HEADING_COMPONENT
-   }
-   
-   // SimpleAtlas Simulation calls this helper constructor
-   public DRCFlatGroundWalkingTrack(RobotInitialSetup<HumanoidFloatingRootJointRobot> robotInitialSetup, DRCGuiInitialSetup guiInitialSetup, DRCSCSInitialSetup scsInitialSetup,
-                                    boolean useVelocityAndHeadingScript, boolean cheatWithGroundHeightAtForFootstep, DRCRobotModel model,
-                                    HeadingAndVelocityEvaluationScriptParameters walkingScriptParameters, HighLevelControllerStateFactory customControllerState)
-   {
-      this(robotInitialSetup, guiInitialSetup, scsInitialSetup, useVelocityAndHeadingScript, cheatWithGroundHeightAtForFootstep, model,
-           WalkingProvider.VELOCITY_HEADING_COMPONENT, walkingScriptParameters, null, customControllerState); // should always be committed as VELOCITY_HEADING_COMPONENT
-   }
-
-   public DRCFlatGroundWalkingTrack(RobotInitialSetup<HumanoidFloatingRootJointRobot> robotInitialSetup, DRCGuiInitialSetup guiInitialSetup,
-                                    DRCSCSInitialSetup scsInitialSetup, boolean useVelocityAndHeadingScript, boolean cheatWithGroundHeightAtForFootstep,
-                                    DRCRobotModel model, PelvisPoseCorrectionCommunicatorInterface externalPelvisCorrectorSubscriber)
-   {
-      this(robotInitialSetup, guiInitialSetup, scsInitialSetup, useVelocityAndHeadingScript, cheatWithGroundHeightAtForFootstep, model,
-           WalkingProvider.VELOCITY_HEADING_COMPONENT, new HeadingAndVelocityEvaluationScriptParameters(), externalPelvisCorrectorSubscriber, null);
-   }
-
-   public DRCFlatGroundWalkingTrack(RobotInitialSetup<HumanoidFloatingRootJointRobot> robotInitialSetup, DRCGuiInitialSetup guiInitialSetup,
-                                    DRCSCSInitialSetup scsInitialSetup, boolean useVelocityAndHeadingScript, boolean cheatWithGroundHeightAtForFootstep,
-                                    DRCRobotModel model, WalkingProvider walkingProvider, HeadingAndVelocityEvaluationScriptParameters walkingScriptParameters)
-   {
-      this(robotInitialSetup, guiInitialSetup, scsInitialSetup, useVelocityAndHeadingScript, cheatWithGroundHeightAtForFootstep, model, walkingProvider,
-           walkingScriptParameters, null, null);
-   }
-
-
-   private DRCFlatGroundWalkingTrack(RobotInitialSetup<HumanoidFloatingRootJointRobot> robotInitialSetup, DRCGuiInitialSetup guiInitialSetup,
-                                    DRCSCSInitialSetup scsInitialSetup, boolean useVelocityAndHeadingScript, boolean cheatWithGroundHeightAtForFootstep,
-                                    DRCRobotModel model, WalkingProvider walkingProvider, HeadingAndVelocityEvaluationScriptParameters walkingScriptParameters,
-                                    PelvisPoseCorrectionCommunicatorInterface externalPelvisCorrectorSubscriber, HighLevelControllerStateFactory customControllerStateFactory)
-   {
-      //    scsInitialSetup = new DRCSCSInitialSetup(TerrainType.FLAT);
-
-      double dt = scsInitialSetup.getDT();
-      int recordFrequency = (int) Math.round(model.getControllerDT() / dt);
-      if (recordFrequency < 1)
-         recordFrequency = 1;
-      scsInitialSetup.setRecordFrequency(recordFrequency);
-
-      HighLevelControllerParameters highLevelControllerParameters = model.getHighLevelControllerParameters();
-      WalkingControllerParameters walkingControllerParameters = model.getWalkingControllerParameters();
-      CoPTrajectoryParameters copTrajectoryParameters = model.getCoPTrajectoryParameters();
-      SplitFractionCalculatorParametersReadOnly splitFractionParameters = model.getSplitFractionCalculatorParameters();
-      HumanoidRobotSensorInformation sensorInformation = model.getSensorInformation();
-      SideDependentList<String> feetForceSensorNames = sensorInformation.getFeetForceSensorNames();
-      SideDependentList<String> wristForceSensorNames = sensorInformation.getWristForceSensorNames();
-
-      RobotContactPointParameters<RobotSide> contactPointParameters = model.getContactPointParameters();
-      ArrayList<String> additionalContactRigidBodyNames = contactPointParameters.getAdditionalContactRigidBodyNames();
-      ArrayList<String> additionalContactNames = contactPointParameters.getAdditionalContactNames();
-      ArrayList<RigidBodyTransform> additionalContactTransforms = contactPointParameters.getAdditionalContactTransforms();
-
-      ContactableBodiesFactory<RobotSide> contactableBodiesFactory = new ContactableBodiesFactory<>();
-      contactableBodiesFactory.setFootContactPoints(contactPointParameters.getFootContactPoints());
-      contactableBodiesFactory.setToeContactParameters(contactPointParameters.getControllerToeContactPoints(), contactPointParameters.getControllerToeContactLines());
-      for (int i = 0; i < contactPointParameters.getAdditionalContactNames().size(); i++)
-         contactableBodiesFactory.addAdditionalContactPoint(additionalContactRigidBodyNames.get(i), additionalContactNames.get(i), additionalContactTransforms.get(i));
-
-      HighLevelHumanoidControllerFactory controllerFactory = new HighLevelHumanoidControllerFactory(contactableBodiesFactory,
-                                                                                                    feetForceSensorNames,
-                                                                                                    wristForceSensorNames,
-                                                                                                    highLevelControllerParameters,
-                                                                                                    walkingControllerParameters,
-                                                                                                    copTrajectoryParameters,
-                                                                                                    splitFractionParameters);
-      setupHighLevelStates(controllerFactory, feetForceSensorNames, highLevelControllerParameters.getFallbackControllerState());
-      controllerFactory.createControllerNetworkSubscriber(model.getSimpleRobotName(), realtimeROS2Node);
-      if (customControllerStateFactory != null)
-         controllerFactory.addCustomControlState(customControllerStateFactory);
-
-      FootstepAdjustment footstepAdjustment = null;
-      if (cheatWithGroundHeightAtForFootstep)
-      {
-         footstepAdjustment = new HeightMapBasedFootstepAdjustment(scsInitialSetup.getHeightMap());
-      }
-
-      AvatarSimulationFactory avatarSimulationFactory = new AvatarSimulationFactory();
-      avatarSimulationFactory.setRobotModel(model);
-      avatarSimulationFactory.setShapeCollisionSettings(model.getShapeCollisionSettings());
-      avatarSimulationFactory.setHighLevelHumanoidControllerFactory(controllerFactory);
-      avatarSimulationFactory.setCommonAvatarEnvironment(null);
-      avatarSimulationFactory.setRobotInitialSetup(robotInitialSetup);
-      avatarSimulationFactory.setSCSInitialSetup(scsInitialSetup);
-      avatarSimulationFactory.setGuiInitialSetup(guiInitialSetup);
-      avatarSimulationFactory.setRealtimeROS2Node(realtimeROS2Node);
-      avatarSimulationFactory.setCreateYoVariableServer(createYoVariableServer);
-      avatarSimulationFactory.setComponentBasedFootstepDataMessageGeneratorParameters(useVelocityAndHeadingScript, footstepAdjustment, walkingScriptParameters);
-
-      if (externalPelvisCorrectorSubscriber != null)
-         avatarSimulationFactory.setExternalPelvisCorrectorSubscriber(externalPelvisCorrectorSubscriber);
-
-      avatarSimulation = avatarSimulationFactory.createAvatarSimulation();
-      initialize();
-
-      if (robotInitialSetup.supportsReset())
-      {
-         JButton resetButton = new JButton("Reset Robot");
-         resetButton.addActionListener(e -> avatarSimulation.resetRobot());
-         avatarSimulation.getSimulationConstructionSet().addButton(resetButton);
-      }
-
-      ((YoBoolean) avatarSimulation.getSimulationConstructionSet().getRootRegistry().findVariable("ignoreWalkInputProviderCSG")).set(true);
-
-      avatarSimulation.start();
-   }
-
-   /**
-    * used to inject anything you need into scs before the sim starts
-    */
-   public void initialize()
-   {
-
-   }
-
-   public void setupHighLevelStates(HighLevelHumanoidControllerFactory controllerFactory, SideDependentList<String> feetForceSensorNames,
-                                    HighLevelControllerName fallbackControllerState)
-   {
-      controllerFactory.useDefaultDoNothingControlState();
-      controllerFactory.useDefaultWalkingControlState();
-
-      controllerFactory.addRequestableTransition(DO_NOTHING_BEHAVIOR, WALKING);
-      controllerFactory.addRequestableTransition(WALKING, DO_NOTHING_BEHAVIOR);
-
-      controllerFactory.addControllerFailureTransition(DO_NOTHING_BEHAVIOR, fallbackControllerState);
-      controllerFactory.addControllerFailureTransition(WALKING, fallbackControllerState);
-
-      controllerFactory.setInitialState(HighLevelControllerName.WALKING);
-   }
-
-   public void attachControllerFailureListener(ControllerFailureListener listener)
-   {
-      avatarSimulation.getHighLevelHumanoidControllerFactory().attachControllerFailureListener(listener);
-   }
-
-   public SimulationConstructionSet getSimulationConstructionSet()
-   {
-      return avatarSimulation.getSimulationConstructionSet();
-   }
-
-   public AvatarSimulation getAvatarSimulation()
-   {
-      return avatarSimulation;
-   }
-
-   public void destroySimulation()
-   {
-      if (avatarSimulation != null)
-      {
-         avatarSimulation.dispose();
-      }
-   }
-}
